Rename decoding testing utils
